removing strict dependency on 'git' import
When git executable is not present, renag should behave in the same way as it does in a normal directory.

But since it might lead to unexpected result, we warn user about the fact that executable is missing without checking whether it was intended to be a git repo or not.